Icons for some astronomy apps.
Modified "celestia.svg" to differentiate it from "stellarium.svg".

Celestia is an astronomy simulator that let's you travel inside the whole solar system and beyond, so it seemed fit to have a planet in its icon.

Added "kstars.svg"; the original icon of this app features a telescope, which I have included here.